requirements.txt
google-genai
pyTelegramBotAPI
pillow
python-dotenv
flask